Primeira fase de alteração do projeto para as novas especificações.
- Lista de alterações:
 - Adequação do projeto ao novo modelo de instrução (muitos TADs foram alterados).

OBS: Alterações de extremo risco.

+++ src/inputSystem/instMemory/instMemory.h
@@ -24,6 +24,7 @@
 #include <stdint.h>
 
 #include "../../cpuError.h"
+#include "../../instruction/instruction.h"
 
 #define INSTMEMORY_EALLOC        NULL
 #define INSTMEMORY_EALLOC_MSG    "Falha ao criar memória de instruções."
@@ -31,7 +32,7 @@
 
 typedef struct _instMemory
 {
-    uint32_t *  mem;
+    uint8_t *   mem;
     uint64_t    length;
 } INSTMEMORY;
